dockerfile: split Mount struct to dockerui/types
Will be used in the follow-up commits for implementing "Dockerfile hooks"
(issue 4576)
